Add support for UniversalLinks to launch the ExampleApp or any other CIQ app from the Garmin Connect app
In addition to using URLScheme to launch the CIQ app from the Garmin Connect app, a new method is now supported that uses UniversalLinks. Garmin Connect app with version 5.6 onwards can use UniversalLinks to launch the CIQ app. This version of the SDK accepts the UniversalLinks along with the payload provided which primarily is the list of CIQ supported devices paired through the Garmin Connect app. Please review the updated documentation for additional details.